Models: Adds a new types file for QCA defined types
